+++ packages/goldsky-management/views/decodedDispatch.sql
@@ -1,0 +1,145 @@
+CREATE
+OR REPLACE VIEW "<%= location; %>"."<%= name; %>" AS WITH decoded AS (
+  SELECT
+    dispatch.*,
+    (
+      (
+        ('x' :: text || "substring"(dispatch.message, 3, 8))
+      ) :: bit(32)
+    ) :: integer AS origin_domain_id,
+    (
+      '0x' :: text || ltrim(
+        "substring"(dispatch.message, (3 + 8), 64),
+        '0' :: text
+      )
+    ) AS sender_address,
+    (
+      (
+        (
+          'x' :: text || "substring"(dispatch.message, ((3 + 8) + 64), 8)
+        )
+      ) :: bit(32)
+    ) :: integer AS nonce,
+    (
+      (
+        (
+          'x' :: text || "substring"(dispatch.message, (((3 + 8) + 64) + 8), 8)
+        )
+      ) :: bit(32)
+    ) :: integer AS destination_domain_id,
+    (
+      '0x' :: text || ltrim(
+        "substring"(dispatch.message, ((((3 + 8) + 64) + 8) + 8), 64),
+        '0' :: text
+      )
+    ) AS recipient_address,
+    "substring"(
+      dispatch.message,
+      (((((3 + 8) + 64) + 8) + 8) + 64)
+    ) AS message_body,
+    "substring"(
+      dispatch.message,
+      (((((3 + 8) + 64) + 8) + 8) + 64),
+      72
+    ) AS message__token_body,
+    (
+      (
+        'x' :: text || "substring"(
+          dispatch.message,
+          (((((3 + 8) + 64) + 8) + 8) + 64),
+          8
+        )
+      ) :: bit(32)
+    ) :: bigint AS message__token__domain,
+    "substring"(
+      dispatch.message,
+      ((((((3 + 8) + 64) + 8) + 8) + 64) + 8),
+      64
+    ) AS message__token__id,
+    "substring"(
+      dispatch.message,
+      (((((((3 + 8) + 64) + 8) + 8) + 64) + 8) + 64),
+      194
+    ) AS action_body,
+    "substring"(
+      dispatch.message,
+      (((((((3 + 8) + 64) + 8) + 8) + 64) + 8) + 64),
+      2
+    ) AS message__action__type,
+    (
+      '0x' :: text || ltrim(
+        "substring"(
+          dispatch.message,
+          (
+            (((((((3 + 8) + 64) + 8) + 8) + 64) + 8) + 64) + 2
+          ),
+          64
+        ),
+        '0' :: text
+      )
+    ) AS message__action__to,
+    (
+      (
+        (
+          'x' :: text || ltrim(
+            "substring"(
+              dispatch.message,
+              (
+                (
+                  (((((((3 + 8) + 64) + 8) + 8) + 64) + 8) + 64) + 2
+                ) + 64
+              ),
+              64
+            ),
+            '0' :: text
+          )
+        )
+      ) :: bit(32)
+    ) :: bigint AS message__action__amount,
+    "substring"(
+      dispatch.message,
+      (
+        (
+          (
+            (((((((3 + 8) + 64) + 8) + 8) + 64) + 8) + 64) + 2
+          ) + 64
+        ) + 64
+      ),
+      64
+    ) AS message__action__details_hash
+  FROM
+    <%= dispatch;%>
+)
+
+SELECT
+  decoded.*,
+  CASE
+    length(decoded.message_body)
+    WHEN (133 * 2) THEN 'TRANSFER' :: text
+    ELSE 'OTHER' :: text
+  END AS message_type,
+
+  CASE
+    decoded.origin_domain_id
+    <% for (const domain of domains) { %>
+      WHEN <%= domain.domain %> THEN '<%= domain.name %>' :: text
+    <% } %>
+    ELSE 'OTHER' :: text
+  END AS origin_domain_name,
+
+  CASE
+    decoded.destination_domain_id
+    <% for (const domain of domains) { %>
+      WHEN <%= domain.domain %> THEN '<%= domain.name %>' :: text
+    <% } %>
+    ELSE 'OTHER' :: text
+  END AS destination_domain_name,
+
+  CASE
+    decoded.message__token__domain
+    <% for (const domain of domains) { %>
+      WHEN '<%= domain.domain %>' :: bigint THEN '<%= domain.name %>' :: text
+    <% } %>
+    ELSE 'OTHER' :: text
+  END AS message__token__domain_name
+FROM decoded;