+++ packages/goldsky-management/views/processFailure.sql
@@ -1,0 +1,2 @@
+CREATE
+OR REPLACE VIEW "<%= location; %>"."<%= name; %>" AS SELECT * from <%= process_failure; %>;