+++ packages/local-environment/prepare.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-export $(cat ./.env | xargs)
+export $(cat ./.env | grep -v '#' | xargs)
 pushd ./hardhat
 docker build -t hardhat:latest .
 popd