+++ packages/sdk-bridge/src/backend.ts
@@ -192,6 +192,7 @@ export class GoldSkyBridgeBackend
     f: Partial<MessageFilter>,
     limit?: number,
   ): Promise<GoldSkyBridgeMessage[] | undefined> {
+    const eventsTable = `${this.env}_views_bridge_events`;
     const query = gql`
       query Query(
         $committedRoot: String
@@ -199,7 +200,7 @@ export class GoldSkyBridgeBackend
         $transactionHash: String
         $limit: Int
       ) {
-        bridge_events(
+        ${eventsTable}(
           where: {
             _or: [
               { dispatch_tx: { _eq: $transactionHash } }
@@ -268,7 +269,7 @@ export class GoldSkyBridgeBackend
 
     const response = await request(this.uri, query, filter, this.headers);
 
-    const events: GoldSkyBridgeMessage[] = nulls2undefined(response.events);
+    const events: GoldSkyBridgeMessage[] = nulls2undefined(response[eventsTable]);
 
     if (!events || events.length <= 0) return undefined;